MovieManagementSystem/requirements.txt
Flask==2.0.2
Flask-Login==0.5.0
scikit-surprise==1.1.1
